Improve sidecar message handling (#1057)
* Improve sidecar message handling

Several changes:
  - allow side-passing of a static "context" to the sidecar process to keep
    messages as small as possible
  - clean up the interfaces and make it clearer what is what (ie: there is
    an emitter inside the process and then a sidecar that runs outside)
  - better error checking
  - better handling of shutdown (sidecars will now properly get a shutdown
    message giving them the opportunity to clean up -- this may add a bit
    of latency at the end of the program execution.

Also improved message when plugins don't load in case of an issue with
Metaflow extensions

* Add the possibility to pass additional context to sidecars

This allows more information to be provided after the sidecar is
created but still maintains small message sizes

* Better handling of initial context message for sidecars

In the previous implementation, a failure to send the initial context
message could crash the flow. This change properly handles that message
like any other message (so has no impact on the flow itself). It also
implements better retry policies to send the context. Finally, this change
also improves error handling: if a message fails to send, in the previous
implementation, it was likely that the next successful one would be
an invalid json.

* Improve handling of corner conditions on message send

* Restart sidecar faster when sidecar process dies

* Properly send context after sidecar restart

* Remove the specific context from monitors/loggers

Subclasses can now handle contexts directly in the subclass.

Refactoring also cleaned up what a sidecar and a sidecar worker
do by refactoring some of the code out.

* Typo: fix names

* Minor typos for sidecar changes

* Clean up sidecars

* More cleanups for sidecars

* Addressed final comments for sidecars